requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pymysql==1.1.0
python-dotenv==1.0.0
boto3==1.28.85
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
APScheduler==3.10.4
